revert: use clap 3

+++ src/args.rs
@@ -1,7 +1,7 @@
 #![allow(clippy::module_name_repetitions)]
 
-use crate::{fl, operations::SearchBy};
-use clap::{ArgAction, Parser, Subcommand, ValueHint};
+use crate::operations::SearchBy;
+use clap::{Parser, Subcommand, ValueHint};
 
 static VERSION: &str = concat!(
     env!("CARGO_PKG_VERSION"),
@@ -13,54 +13,67 @@ static VERSION: &str = concat!(
 #[derive(Debug, Clone, Parser)]
 #[clap(bin_name = "ame", name = "Amethyst", version = VERSION, about = env ! ("CARGO_PKG_DESCRIPTION"), infer_subcommands = true)]
 pub struct Args {
-    #[command(subcommand)]
+    #[clap(subcommand)]
     pub subcommand: Option<Operation>,
 
-    #[arg(long, short, action = ArgAction::Count, global = true, help = fl!("verbose"))]
-    pub verbose: u8,
+    /// Sets the level of verbosity
+    #[clap(long, short, parse(from_occurrences), global = true)]
+    pub verbose: usize,
 
-    #[arg(long = "noconfirm", global = true, help = fl!("no-confirm"))]
+    /// Complete operation without prompting user
+    #[clap(long = "noconfirm", global = true)]
     pub no_confirm: bool,
 
+    /// Make some commands have less output (only clean, upgrade, and install are supported)
+    #[clap(long, short, global = true)]
     // not gonna lie the only reason this option is here is because
     // i was trying to figure out if paccache had a --noconfirm option.
     // so basically, it doesn't, but hey, we might as well have it here
     // anyways as some pacman commands can have a --quiet flag passed
     // to them.
-    #[arg(long, short, global = true, help = fl!("quiet"))]
     pub quiet: bool,
 
-    #[arg(long = "sudoloop", global = true, help = fl!("sudoloop"))]
+    /// Loops sudo in the background to ensure it doesn't time out during long builds
+    #[clap(long = "sudoloop", global = true)]
     pub sudoloop: bool,
 
-    #[arg(long, short, global = true, value_hint = ValueHint::DirPath, help = fl!("cachedir"))]
+    /// Sets a custom AUR clone and build directory for the specified operation
+    #[clap(long, short, global = true, value_hint = ValueHint::DirPath)]
     pub cachedir: Option<String>,
 }
 
 #[derive(Debug, Clone, Subcommand)]
 pub enum Operation {
-    #[command(bin_name = "ame", name = "sync", aliases = & [ "-S" ], visible_aliases = & ["install", "i"], short_flag = 'S', about = fl!("install"))]
+    /// Installs or searches for a package in either the AUR or the Pacman-defined repositories
+    #[clap(bin_name = "ame", name = "sync", aliases = & [ "-S" ], visible_aliases = & ["install", "i"], short_flag = 'S')]
     Install(InstallArgs),
 
-    #[command(bin_name = "ame", name = "remove", visible_aliases = & ["rm"], short_flag = 'R', about = fl!("remove"))]
+    /// Removes a previously installed package
+    #[clap(bin_name = "ame", name = "remove", visible_aliases = & ["rm"], short_flag = 'R')]
     Remove(RemoveArgs),
 
-    #[command(bin_name = "ame", name = "search", about = fl!("search"))]
+    /// Searches for packages matching a provided pattern in the AUR/repos [aliases: -Ss]
+    #[clap(bin_name = "ame", name = "search")]
     Search(InstallArgs),
 
-    #[command(bin_name = "ame", name = "query", short_flag = 'Q', about = fl!("query"))]
+    /// Queries installed packages
+    #[clap(bin_name = "ame", name = "query", short_flag = 'Q')]
     Query(QueryArgs),
 
-    #[command(bin_name = "ame", name = "upgrade", visible_aliases = & ["-Syu"], about = fl!("upgrade"))]
+    /// Upgrades locally installed packages to their latest versions (Default)
+    #[clap(bin_name = "ame", name = "upgrade", visible_aliases = & ["-Syu"])]
     Upgrade(UpgradeArgs),
 
-    #[command(bin_name = "ame", name = "gencomp", short_flag = 'G', about = fl!("gencomp"))]
+    /// Generates shell completions for supported shells (bash, fish, elvish, pwsh)
+    #[clap(bin_name = "ame", name = "gencomp", short_flag = 'G')]
     GenComp(GenCompArgs),
 
-    #[command(bin_name = "ame", name = "clean", short_flag = 'C', about = fl!("clean"))]
+    /// Removes all orphaned packages
+    #[clap(bin_name = "ame", name = "clean", short_flag = 'C')]
     Clean,
 
-    #[command(bin_name = "ame", name = "diff", short_flag = 'd', about = fl!("diff"))]
+    /// Runs pacdiff
+    #[clap(bin_name = "ame", name = "diff", short_flag = 'd')]
     Diff,
 }
 
@@ -72,51 +85,63 @@ impl Default for Operation {
 
 #[derive(Default, Debug, Clone, Parser)]
 pub struct InstallArgs {
-    #[arg(required = true, help = fl!("install-packages"))]
+    /// The name of the package(s) to install or search for
+    #[clap(required = true)]
     pub packages: Vec<String>,
 
-    #[arg(long, short, help = fl!("install-aur"))]
+    /// Operate only on AUR packages
+    #[clap(long, short)]
     pub aur: bool,
 
-    #[arg(long, short, help = fl!("install-repo"))]
+    /// Operate only on repo packages
+    #[clap(long, short)]
     pub repo: bool,
 
-    #[arg(hide = true, short = 's', help = fl!("install-search"))]
+    /// Search packages for a given pattern instead of installing
+    #[clap(hidden = true, short = 's')]
     pub search: bool,
 
-    #[arg(long, short, help = fl!("install-by"))]
+    /// Searches by a specific field
+    #[clap(long, short)]
     pub by: Option<SearchBy>,
 }
 
 #[derive(Default, Debug, Clone, Parser)]
 pub struct RemoveArgs {
-    #[arg(required = true, help = fl!("remove-packages"))]
+    /// The name of the package(s) to remove
+    #[clap(required = true)]
     pub packages: Vec<String>,
 }
 
 #[derive(Default, Debug, Clone, Parser)]
 pub struct QueryArgs {
-    #[arg(long, short, help = fl!("query-aur"))]
+    /// Lists AUR/foreign packages [-Qa, -Qm]
+    #[clap(long, short)]
     pub aur: bool,
 
-    #[arg(long, short, help = fl!("query-repo"))]
+    /// Lists repo/native packages [-Qr, -Qn]
+    #[clap(long, short)]
     pub repo: bool,
 
-    #[arg(long, short, help = fl!("query-info"))]
+    /// Get information about a specific package
+    #[clap(long, short)]
     pub info: Option<String>,
 }
 
 #[derive(Default, Debug, Clone, Parser)]
 pub struct UpgradeArgs {
-    #[arg(long, short, help = fl!("upgrade-repo"))]
+    /// Upgrades only repo/native packages
+    #[clap(long, short)]
     pub repo: bool,
 
-    #[arg(long, short, help = fl!("upgrade-aur"))]
+    /// Upgrades only from the AUR
+    #[clap(long, short)]
     pub aur: bool,
 }
 
 #[derive(Default, Debug, Clone, Parser)]
 pub struct GenCompArgs {
-    #[arg(required = true, help = fl!("gencomp-shell"))]
+    /// The shell to generate completions for (bash, fish, elvish, pwsh, fig)
+    #[clap(required = true)]
     pub shell: String,
 }
